- Додано можливість редагувати співробітників через адмін - панель - Виправлено стилі для каруселі товарів - Частково налаштовані дозволи в адмін-панелі по ролям

+++ src/app/AdminPanel/tsx/FrameShop.tsx
@@ -12,12 +12,12 @@ import { toast } from 'react-toastify';
 import '../css/ShopsFrame.css'; // Імпортуємо CSS файл
 import themeFrame from './ThemeFrame';
 
-export default function FrameShop() {
+export default function FrameShop({ rolePermissions }) {
   const [activeTab, setActiveTab] = useQueryState("at", { defaultValue: "products", scroll: false, history: "push", shallow: true });
   const { data: data = [], isLoading: loading } = useShops({
     SearchParameter: "Query",
     PageNumber: 1,
-    PageSize: 150
+    PageSize: 150,
   });
 
   const { mutateAsync: deleteShop } = useDeleteShop();
@@ -69,7 +69,7 @@ export default function FrameShop() {
   };
 
   // Створюємо масив колонок з перекладеними назвами
-  const columns = [
+  let columns = [
     { field: 'id', headerName: 'ID', flex: 0.3, minWidth: 50 },
     { field: 'name', headerName: 'Назва магазину', flex: 1, minWidth: 200 },
     { field: 'state', headerName: 'Область', flex: 1, minWidth: 150 },
@@ -79,31 +79,48 @@ export default function FrameShop() {
     { field: 'postalCode', headerName: 'Поштовий індекс', flex: 1, minWidth: 150 },
     {
       field: 'executedOrdersSum',
-      headerName: 'Заг. сума товарів',
+      headerName: 'Заг. сума виконаних замовлень',
       flex: 0.5,
       cellClassName: 'text-right',
-      renderCell: (params) => formatCurrency(params.value),
+      renderCell: (params) => {
+        if (rolePermissions.canReadShopExecutedOrdersSum(params.row.id)) {
+          return formatCurrency(params.value)
+        }
+        else {
+          return null;
+        }
+      },
     },
     {
       field: 'actions',
       headerName: 'Дії',
-      flex: 0.5,
-      minWidth: 175,
-      renderCell: (params) => (
-        <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', gap: "5px", height: "100%" }}>
-
-          <Button sx={{ minWidth: "10px", padding: 0, color: "#00AAAD" }} title='Редагувати' variant="outlined" onClick={() => handleEdit(params.row)}>
-            <EditIcon />
-          </Button>
-          <Button sx={{ minWidth: "10px", padding: 0, color: '#be0f0f' }} title='Видалити' variant="outlined" color="secondary" onClick={() => handleDelete(params.row)}>
-
-            <DeleteIcon />
-          </Button>
-        </Box>
-      ),
+      flex: 0,
+      width: 75,
+      renderCell: (params) => {
+        if (rolePermissions.IsFrameShops_Button_EditShop_Available || rolePermissions.IsFrameShops_Button_DeleteShop_Available) {
+          return <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', gap: "5px", height: "100%" }}>
+            {rolePermissions.IsFrameShops_Button_EditShop_Available && rolePermissions.canEditShopAsAdmin(params.row.id) &&
+              <Button sx={{ minWidth: "10px", padding: 0, color: "#00AAAD" }} title='Редагувати' variant="outlined" onClick={() => handleEdit(params.row)}>
+                <EditIcon />
+              </Button>}
+            {rolePermissions.IsFrameShops_Button_DeleteShop_Available &&
+              <Button sx={{ minWidth: "10px", padding: 0, color: '#be0f0f' }} title='Видалити' variant="outlined" color="secondary" onClick={() => handleDelete(params.row)}>
+                <DeleteIcon />
+              </Button>}
+          </Box>
+        }
+        else {
+          return null;
+        }
+      },
     },
   ];
-
+  if (!rolePermissions.IsFrameShops_Cell_ExecutedOrdersSum_Available) {
+    columns = columns.filter((column) => column.field !== 'executedOrdersSum');
+  }
+  if (!rolePermissions.IsFrameShops_Cell_Actions_Available) {
+    columns = columns.filter((column) => column.field !== 'actions');
+  }
   const handleEdit = (row) => {
     setShopId(row.id);
     console.log("row.id :", row.id)
@@ -124,11 +141,11 @@ export default function FrameShop() {
             value={searchTerm}
             onChange={(e) => setSearchTerm(e.target.value)} // Оновлюємо стан для швидкого пошуку
           />
-          <Button sx={{ backgroundColor: "#00AAAD" }} variant="contained" onClick={() => {
+          {rolePermissions.IsFrameShops_Button_AddShop_Available && <Button sx={{ backgroundColor: "#00AAAD" }} variant="contained" onClick={() => {
             setShopId(0); setActiveTab('addNewShop')
           }}>
             Додати
-          </Button>
+          </Button>}
         </Box>
         <Box sx={{ overflowX: 'auto' }} height="80vh"> {/* Додаємо прокрутку при переповненні */}
           <DataGrid
